Prestiges amount now affects XP. We get 1 after every fifth level starting from 15 from now on. Level required to prestige is increased to 10 again. Added adjustment for phones for CoresPage

+++ DopamineXP/Models/DopamineTask.cs
@@ -32,10 +32,10 @@ public class DopamineTask
     public int Cores { get; set; } = 0;
     public int AmountOfPrestiges { get; set; } = 0;
     
-    public int Shards { get; set; } = 0;
+    public double Shards { get; set; } = 0;
     public DateTime LastShardUpdate { get; set; } = DateTime.Now;
 
-    public int FirstGeneratorsAmount { get; set; } = 0;
+    public double FirstGeneratorsAmount { get; set; } = 0;
     
     
     public string GetRgba(double alpha)
